Fix brainfart: no locking in session
At this time, there's only 1 service thread, so any operations not on a
handler chain can simply be posted to the io service.

In fact, if synchronization is required later (e.g. because there are
mormore than 1 service threads running), then we should create a strand
for all operations on a `connection`.

See e.g.
https://stackoverflow.com/questions/48507883/do-i-need-to-implement-blocking-when-using-boostasio/48523707#48523707
which gradually introduces more sophisticated Asio synchronization as
required